Much much  happier
Last push was a fluke. This code correctly updates the progress of the ever-expanding MAX_SMA_DAY. Thinking that it might be time to create a pull request.

Last few remaining items on the checklist before I feel good doing a pull to main. ..

we're currently not calculating pairs in max function where SMA 1 is larger than SMA 2.

i'll address more. stuff like, needing top pair print at end of console... and existing_max_sma_day not updating with new MAX_SMA_DAY if MAX_SMA_DAY stays the same.

and we might be doubling up on compuations... where short results are simply the inverse of buy_results...

and total_trading_days is probably not working at the moment to update the row to reflect the latest day increase.