Successful update!!
Goal achieved. All trading pairs now considered regardless of size of SMA1 and SMA2.

Still need to work out some bugs. New trading days not included in calculations for tickers with existing pkl files.

Will also have to re-visit possible parallel computations that are occurring -- it offers some additional confidence in the calculations but it could be doubling the precompute times. Basically, short results should not have to be calculated separately, because they are simply the inverse of the buy pairs.