More accurate now
Secondary ticker metrics were not matching up as intended. Also updated aspects of the Dynamic Master Trading Strategy Section that were not reporting correct figures for some of the averages. A loss is now defined as any day with a buy or a sell signal that results in anything less than or equal to zero for our returns the next day.