Error Handling and Other
Changed timezone to EST.

Got rid of ticker clock because that data is not available from yfinance.

Attempted to create exclusions for crypto (needs further testing).

Code does not currently account for holidays... variations.. like stocks being open and bonds being closed... it may try to create a new day when there is no new day close data.

I am not confident that the NoneType bug is gone but it is related to the off-by-one issue with a ticker history being one day too long or short.