Server Timeout in the Automation Section SHOULD be fixed.
Added some improved features like a progress bar and better error handling. Some bugs are still present that I haven't figured out how to fix like one where when you are on page two of the table results in this automation section, clicking on a cell in the category column sends the build to the multi-primary that is associated with the corresponding row from page one of the table -- instead of sending the build from the relevant cell on the other page (TLDR, basically any other table page than page one doesn't send the correct build).

Set MAX primary tickers in this section to 18 -- any more than that and we are talking about days of computational work.