Adjusted metrics to match across the board
Numbers matching up correctly now.

ALSO -- Made a small adjustment with how top pair ties are calculated -- we are now defaulting to slightly different way of dealing with top pair ties but one that I feel lines up with the rules of probability better than it's predecessor. Here is more:

Explanation of Changes:

a. The specific problem addressed:
Currently, np.argmax() returns the first occurrence of the maximum value. In a tie, this picks the “front” of the list. By reversing the array before using argmax(), we prioritize the latter occurrences in case of ties, effectively flipping the approach.

b. How the changes solve the problem:
By using [::-1] and then adjusting the index, we ensure that ties are resolved starting from the end of the array rather than the beginning.