Multi-Primary Signal Aggregator Implementation
GREAT WORK!

Really great work.

Only a few more stops but that was a huge breakthrough.

The next step in our confluence journey will be to take our analysis that we generate and then to have a way to automatically apply it to our "Analyze Confluence" section -- as it stands, the user only has the ability to "Analyze Confluence" for one ticker at a time. The goal is to allow for our results to all be applied in the analyze confluence section, so that the user can see what all of the current signals are saying, how they all stack up etc. Does that make sense? Game this out with me before recommending changes.